requirements.txt
python-twitter
numpy
scipy
scikit-learn
matplotlib
pandas
selenium
pypng
pillow